% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaff_boxplot.R
\name{scaff_boxplot}
\alias{scaff_boxplot}
\title{Creating a Boxplot by Group}
\usage{
scaff_boxplot(data, x, y, xname = NULL, yname = NULL)
}
\arguments{
\item{data}{dataset to use for plot.}

\item{x}{Categorical Predictor.}

\item{y}{Outcome of Interest.}

\item{xname}{(Optional) Name of Categorical Predictor you would like displayed in the plot. If not specified, the name of data$x is used instead.}

\item{xyname}{(Optional) Name of Outcome of Interest you would like displayed in the plot. If not specified, the name of data$y is used instead.}
}
\description{
This allows you to create a plot where there is a boxplot for each category.
}
\examples{
scaff_boxplot(data = iris,x = Species ,y = Sepal.Length, xname = "Iris Species", yname = "Sepal Length")
}
