% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaff_diagnostics.R
\name{scaff_diagnostics}
\alias{scaff_diagnostics}
\title{Creating a Table of Summary Statistics}
\usage{
scaff_diagnostics(LMinput, globaloutput = FALSE)
}
\arguments{
\item{LMinput}{Object created by linear model fitting function to be passed into diagnostic plots}

\item{globaloutput}{An optional logical indicating whether the ggplot objects should be output to global environment as a list}
}
\description{
This function takes an input of a linear object model and uses ggplot to output typical model diagnostic plots.
}
\examples{
lmOut <- lm(Sepal.Length ~ Species, data = iris)
scaff_diagnostics(lmOut)
}
