% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaff_pvnot.R
\name{scaff_pvnot}
\alias{scaff_pvnot}
\title{Rounding P-Values}
\usage{
scaff_pvnot(pvalue)
}
\arguments{
\item{pvalue}{Value that need to be changed to appropriate notation.}
}
\description{
This function allows you to print any value that is less than 0.001 as '<0.001' instead of using long scientific notation.
}
\examples{
pvnot(9.8E-9)
}
