% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaff_sumtable.R
\name{scaff_sumtable}
\alias{scaff_sumtable}
\title{Creating a Table of Summary Statistics}
\usage{
scaff_sumtable(data, x, y)
}
\arguments{
\item{data}{dataframe passed into function}

\item{x}{Categorical grouping variable}

\item{y}{Outcome of Interest}
}
\description{
This function takes three inputs: a dataframe, a grouping variable x, and an outcome y, and returns a table of summary statistics grouped by the x variable. The table contains commonly-requested parametric and nonparametric summary statistics, and the results can be passed to a formatting command (such as kable or flextable).
}
\examples{
scaff_sumtable(data = iris, x = Sepal.Length, y = Species)
}
